+++ freediscovery/__main__.py
@@ -12,7 +12,7 @@ from .server import fd_app
 from freediscovery.engine.cli import (_query_yes_no, _TeeLogger,
                                       _number_of_workers)
 from .text import FeatureVectorizer
-from .pipeline import PipelineFinder
+from freediscovery.engine.pipeline import PipelineFinder
 from .datasets import IR_DATASETS, load_dataset
 from ._version import __version__
