+++ freediscovery/base.py
@@ -1,20 +1,12 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
 import os
-from collections import OrderedDict
 
-import numpy as np
 from sklearn.externals import joblib
 
 from .text import FeatureVectorizer
-from .pipeline import PipelineFinder
-from .exceptions import ModelNotFound
-from .exceptions import (DatasetNotFound, InitException, NotFound, WrongParameter)
+from freediscovery.engine.pipeline import PipelineFinder
+from freediscovery.exceptions import WrongParameter
 
 
 class RankerMixin(object):
