+++ freediscovery/stop_words.py
@@ -1,16 +1,11 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
 import os
 import os.path
 
 from sklearn.externals.joblib import dump, load
 
-from .pipeline import PipelineFinder
+from freediscovery.engine.pipeline import PipelineFinder
 
 CUSTOM_STOP_WORDS = ['com']
